requirements.txt
amqp==5.0.9
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
camelot-py==0.10.1
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==36.0.1
dateparser==1.1.0
Django==4.0
et-xmlfile==1.1.0
ghostscript==0.7
greenlet==1.1.2
idna==3.3
kombu==5.2.2
numpy==1.21.5
opencv-python==4.5.5.62
openpyxl==3.0.9
pandas==1.3.5
pdfminer.six==20211012
prompt-toolkit==3.0.24
psycopg2==2.9.3
pycparser==2.21
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.2
tabulate==0.8.9
tzdata==2021.5
tzlocal==4.1
unicode==2.8
Unidecode==1.3.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
